ignore AZ changes to RDS cluster (as per TF docs) (#39)
* ignore AZ changes to RDS cluster (as per TF docs)
Unfortunately, TF still does not support variables in lifecycle blocks. However, it is my opinion that the aws_rds_cluster block is extremely unlikely to have just the availability_zones changed as this is likely to also impact the subnets too, and not using all available AZs is an edge case anyway. If it were to happen, then users could work around the issue with taint/refresh commands.

Signed-off-by: Daniel.Hill <daniel.hill@engineering.digital.dwp.gov.uk>